+++ Tests/Integration/DSCResources/SqlServer.config.ps1
@@ -25,12 +25,16 @@ configuration SqlServer_config
         $SkipRuleType,
 
         [Parameter()]
+        [string[]]
+        $SkipRuleSeverity,
+
+        [Parameter()]
         [hashtable]
         $Exception,
 
         [Parameter()]
-        [hashtable]
-        $BackwardCompatibilityException,
+        [string[]]
+        $ResourceParameters,
 
         [Parameter()]
         [object]
@@ -42,44 +46,17 @@ configuration SqlServer_config
 
     Node localhost
     {
-        & ([scriptblock]::Create("
-        SqlServer Instance
-        {
-            SqlVersion = $TechnologyVersion
-            SqlRole = '$TechnologyRole'
-            StigVersion = '$StigVersion'
-            ServerInstance = 'TestServer'
-            $(if ($OrgSettings -is [hashtable])
-            {
-                "Orgsettings = @{`n$($OrgSettings.Keys |
-                    ForEach-Object {"'{0}' = {1}{2} = '{3}'{4}`n" -f
-                        $PSItem, '@{', $($OrgSettings[$PSItem].Keys), $($OrgSettings[$PSItem][$OrgSettings[$PSItem].Keys]), '}'})}"
-            }
-            elseif ($null -ne $OrgSettings)
-            {
-                "Orgsettings = '$OrgSettings'"
-            })
-            $(if ($null -ne $Exception)
-            {
-                "Exception = @{`n$($Exception.Keys |
-                    ForEach-Object {"'{0}' = {1}{2} = '{3}'{4}`n" -f
-                        $PSItem, '@{', $($Exception[$PSItem].Keys), $($Exception[$PSItem][$Exception[$PSItem].Keys]), '}'})}"
-            })
-            $(if ($null -ne $BackwardCompatibilityException)
-            {
-                "Exception = @{`n$($BackwardCompatibilityException.Keys |
-                    ForEach-Object {"'{0}' = {1}`n" -f $PSItem, $BackwardCompatibilityException[$PSItem]})}"
-            })
-            $(if ($null -ne $SkipRule)
-            {
-                "SkipRule = @($( ($SkipRule | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-            }
-            if ($null -ne $SkipRuleType)
-            {
-                " SkipRuleType = @($( ($SkipRuleType | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-            })
-        }")
-        )
+        $psboundParams = $PSBoundParameters
+        $psboundParams.SqlVersion = $psboundParams['TechnologyVersion']
+        $psboundParams.SqlRole = $psboundParams['TechnologyRole']
+        $psboundParams.ServerInstance = 'TestServer'
+        $psboundParams.Remove('TechnologyRole')
+        $psboundParams.Remove('ConfigurationData')
+        $psboundParams.Remove('TechnologyVersion')
+
+        $resourceParamString = New-ResourceParameterString -ResourceParameters $ResourceParameters -PSBoundParams $psboundParams
+        $resourceScriptBlockString = New-ResourceString -ResourceParameterString $resourceParamString -ResourceName SqlServer
+        & ([scriptblock]::Create($resourceScriptBlockString))
     }
 }
 
@@ -122,31 +99,17 @@ configuration SqlServerDatabase_config
 
     Node localhost
     {
-        & ([scriptblock]::Create("
-        SqlServer Database
-        {
-            SqlVersion     = '$TechnologyVersion'
-            SqlRole        = '$TechnologyRole'
-            StigVersion    = '$StigVersion'
-            ServerInstance = 'TestServer'
-            Database       = 'TestDataBase'
-            $(if ($null -ne $OrgSettings)
-            {
-                "Orgsettings = '$OrgSettings'"
-            })
-            $(if ($null -ne $Exception)
-            {
-                "Exception = @{$( ($Exception | ForEach-Object {"'$PSItem'= @{'SetScript'='TestScript'}"}) -join "`n" )}"
-            })
-            $(if ($null -ne $SkipRule)
-            {
-                "SkipRule = @($( ($SkipRule | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-            }
-            if ($null -ne $SkipRuleType)
-            {
-                "SkipRuleType = @($( ($SkipRuleType | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
-            })
-        }")
-        )
+        $psboundParams = $PSBoundParameters
+        $psboundParams.SqlVersion = $psboundParams['TechnologyVersion']
+        $psboundParams.SqlRole = $psboundParams['TechnologyRole']
+        $psboundParams.ServerInstance = 'TestServer'
+        $psboundParams.Database = @('TestDataBase','TestDataBase2')
+        $psboundParams.Remove('TechnologyRole')
+        $psboundParams.Remove('ConfigurationData')
+        $psboundParams.Remove('TechnologyVersion')
+
+        $resourceParamString = New-ResourceParameterString -ResourceParameters $ResourceParameters -PSBoundParams $psboundParams
+        $resourceScriptBlockString = New-ResourceString -ResourceParameterString $resourceParamString -ResourceName SqlServer
+        & ([scriptblock]::Create($resourceScriptBlockString))
     }
 }