+++ Tests/Unit/DSCResources/windows.Registry.config.ps1
@@ -8,7 +8,7 @@ configuration Registry_config
 
     Import-Module $moduleRoot\DscResources\helper.psm1 -Force
     Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.2.0
-    Import-DscResource -ModuleName PSDscResources -ModuleVersion 2.10.0.0
+    Import-DscResource -ModuleName PSDSCresources -ModuleVersion 2.12.0.0
 
     Node localhost
     {