+++ source/DSCResources/Vsphere/Vsphere.schema.psm1
@@ -94,12 +94,17 @@ configuration Vsphere
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [string[]]
-        $SkipRuleType
+        $SkipRuleType,
+
+        [Parameter()]
+        [ValidateSet('CAT_I', 'CAT_II', 'CAT_III')]
+        [string[]]
+        $SkipRuleSeverity
     )
 
     ##### BEGIN DO NOT MODIFY #####
     $stig = [STIG]::New('Vsphere', $Version, $StigVersion)
-    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType)
+    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleSeverity)
     ##### END DO NOT MODIFY #####
 
     Import-DscResource -ModuleName Vmware.vSphereDSC -ModuleVersion 2.1.0.58
@@ -112,6 +117,6 @@ configuration Vsphere
     . "$resourcePath\Vsphere.VmHostVssSecurity.ps1"
     . "$resourcePath\Vsphere.VmHostVssPortGroupSecurity.ps1"
 
-    Import-DscResource -ModuleName PSDscResources -ModuleVersion 2.10.0.0
+    Import-DscResource -ModuleName PSDSCresources -ModuleVersion 2.12.0.0
     . "$resourcePath\windows.Script.skip.ps1"
 }